pyproject.toml
[tool.poetry]
name = "podcast-generator"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-genai = "^1.16.1"
streamlit = "^1.45.1"
flask = "^3.1.1"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
